pyproject.toml
[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.coverage.run]
source = ["apps"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/admin.py",
    "manage.py",
]

[tool.coverage.report]
show_missing = true

[project]
name = "education-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "amqp==5.3.1",
    "asgiref==3.11.1",
    "asttokens==3.0.1",
    "billiard==4.2.4",
    "celery==5.6.2",
    "click==8.3.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "coverage==7.13.4",
    "decorator==5.2.1",
    "django==6.0.2",
    "django-debug-toolbar==6.2.0",
    "django-extensions==4.1",
    "django-redis==6.0.0",
    "django-stubs[compatible-mypy]==5.2.9",
    "django-stubs-ext==5.2.9",
    "djangorestframework==3.16.1",
    "djangorestframework-stubs>=3.16.8",
    "executing==2.2.1",
    "gunicorn>=25.1.0",
    "ipython==9.10.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "kombu==5.6.2",
    "matplotlib-inline==0.2.1",
    "mypy>=1.19.1",
    "packaging==26.0",
    "parso==0.8.6",
    "pillow==12.1.1",
    "pip==25.1.1",
    "prompt-toolkit==3.0.52",
    "psycopg2-binary>=2.9.11",
    "pure-eval==0.2.3",
    "pygments==2.19.2",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "redis==7.2.1",
    "six==1.17.0",
    "sqlparse==0.5.5",
    "stack-data==0.6.3",
    "traitlets==5.14.3",
    "types-pyyaml==6.0.12.20250915",
    "typing-extensions==4.15.0",
    "tzdata==2025.3",
    "tzlocal==5.3.1",
    "vine==5.1.0",
    "wcwidth==0.6.0",
    "whitenoise>=6.12.0",
    "ruff>=0.15.5",
]
